fix: satisfy latest query audit checks
